Import and adapt code from Sequencescape
This commit brings across the current code from Sequencescape
and makes the initial attempts to adapt it to support frameworks
other than Rails/ActiveRecord.

Following on from this the Rails specific behaviour (mainly logging
and transactions) will be extracted to an Adapter class, but want
to get the basics flushed out first.