Default yaml files correctly templated
The yaml files did not have the tt extension, so were not being
templated, and the <%= record_key %> was remaining.

This indicates that I really need to get tests in place for the
generators. It will mean adding Rails as a developmental dependency
but that's better than a series of bugs.